requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# test_scenario.py: 8
gpiozero == 1.4.1

# test_runner.py: 10
junit_xml == 1.8

# tests/test_code_style.py: 7
pep8 == 1.7.1

# communication.py: 12
# flasher.py: 9
pyserial == 3.4

# appveyor/appveyor_manager.py: 10
# appveyor/appveyor_test_wait.py: 9
requests == 2.20.1

